+++ NEWS.md
@@ -7,6 +7,7 @@
 - Updated `ReadXenium` to load cell_feature_matrix.h5 when present in favor of the MEX format files
 - Added ability to read Xenium `segmentation_method` directly into `meta.data`
 - Updated `ReadXenium` to load .parquet files using `arrow` instead of .csv.gz files to support XOA 3.0
+- Updated `RunSLSI` to support `BPCells` matrices
 - Fixed `LoadXenium` to accommodate datasets without "Blank Codeword" or "Unassigned Codeword" matrices
 - Fixed `ReadXenium` to properly parse multiple molecular outputs at once ([#8265](https://github.com/satijalab/seurat/issues/8265))
 - Fixed `RunPCA` to avoid converting `BPCells` matrices into dense matrices - significantly reduces the function's memory usage when running on `BPCells` matrices